ping_test: allow StdDev of 0
it is possible to have multiple pings with the
same (or close enough) time that they result in a
standard deviation of 0.000ms (the ping tool will not
report a higher precision of standard deviation).
This is most likely to happen on loopback addresses.

In this test, we calculate the expected standard
deviation, and then only check for the StdDev if
it is an StdDev which is expressible by the ping
utility.